Move payment method state to PaymentMethodProvider
