+++ packages/composite-checkout/src/components/checkout-payment-methods.tsx
@@ -5,6 +5,7 @@ import debugFactory from 'debug';
 import { useCallback, useContext } from 'react';
 import CheckoutContext from '../lib/checkout-context';
 import joinClasses from '../lib/join-classes';
+import { PaymentMethodProviderContext } from '../lib/payment-method-provider-context';
 import { useAvailablePaymentMethodIds } from '../lib/payment-methods';
 import {
 	useAllPaymentMethods,
@@ -39,7 +40,8 @@ export default function CheckoutPaymentMethods( {
 	className?: string;
 } ) {
 	const { __ } = useI18n();
-	const { onPageLoadError, onPaymentMethodChanged } = useContext( CheckoutContext );
+	const { onPageLoadError } = useContext( CheckoutContext );
+	const { onPaymentMethodChanged } = useContext( PaymentMethodProviderContext );
 	const onError = useCallback(
 		( error: Error ) => onPageLoadError?.( 'payment_method_load', error ),
 		[ onPageLoadError ]
