+++ packages/composite-checkout/src/lib/checkout-context.ts
@@ -2,16 +2,8 @@ import { createContext } from 'react';
 import { CheckoutContextInterface } from '../types';
 
 const defaultCheckoutContext: CheckoutContextInterface = {
-	allPaymentMethods: [],
-	disabledPaymentMethodIds: [],
-	setDisabledPaymentMethodIds: noop,
-	paymentMethodId: null,
-	setPaymentMethodId: noop,
 	paymentProcessors: {},
 };
-
-// eslint-disable-next-line @typescript-eslint/no-empty-function
-function noop(): void {}
 
 const CheckoutContext = createContext( defaultCheckoutContext );
